Clean up of code

+++ code/ancient-roman/code/font.h
@@ -5,9 +5,9 @@
 
 extern "C" {
 
-	int GetLetterWidth(u32 letter);
+	u8 GetLetterWidth(u32 letter);
 
-	u8 GetSentenceWidth(const char* text, u32 curIdx, u8* graphic);
+	u16 GetSentenceWidth(const char* text, u32 curIdx, u8* graphic);
 }
 
 #endif
